refactor Translation -> Text
We do not have Translations anymore but work directly with Texts. They
now have a phantom type indicating if the text is static or dynamic. To
make the text dynamic you have to give it a `name` and then should
provide string identifiers for all argument accessors using argString,
etc.. This way, writing a text by hand requires less boilerplate.